+++ eslint.config.js
@@ -47,6 +47,7 @@ const rules = {
   'no-use-before-define': ['error', { functions: false, classes: true, variables: true }],
   'no-var': 0,
   'no-with': 2,
+  'prefer-arrow-callback': ['error', { allowNamedFunctions: true }],
   'prefer-const': ['error', { destructuring: 'all' }],
   'require-atomic-updates': 0,
   'require-await': 2,
@@ -66,7 +67,7 @@ export default [
     languageOptions: {
       parserOptions: {
         sourceType: 'module',
-        ecmaVersion: 2020,
+        ecmaVersion: 2022,
       },
     },
     rules,