+++ test/helpers/setup.js
@@ -1,4 +1,2 @@
-import 'chai/register-expect.js';
-
 process.env.TZ = 'Europe/Stockholm';
 process.env.NODE_ENV = 'test';